# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Trade API
  version: 1.0.0
  description: Sandbox crypto trading API for students. Covers market data, wallets, buy, sell, swap, transactions, KYC, and admin operations.
servers:
  - url: /
    description: Current API host
  - url: http://localhost:4200
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /wallet/portfolio/history:
    get:
      tags: [Wallet]
      summary: Get portfolio chart history
      parameters:
        - in: query
          name: range
          required: false
          schema:
            type: string
            enum: [1D, 1W, 1M, 1Y]
          example: 1M
      responses:
        "200":
          description: Portfolio chart points
          content:
            application/json:
              examples:
                success:
                  value:
                    data:
                      - time: "2026-05-20T12:00:00.000Z"
                        valueUsd: 8942.32
                    meta:
                      count: 30
                      range: 1M
                      latestValueUsd: 9011.44
...
  /wallet/deposit-addresses/{symbol}:
    get:
      tags: [Wallet]
      summary: Get a deposit address and QR payload
      parameters:
        - $ref: "#/components/parameters/Symbol"
      responses:
        "200":
          description: Deposit address
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/DepositAddressResponse"
  /wallet/transactions:
    get:
      tags: [Wallet]
      summary: List user transactions
      parameters:
        - in: query
          name: status
          required: false
          description: Optional transaction status filter.
          schema:
            type: string
            enum: [pending, completed, failed, cancelled]
          example: completed
        - in: query
          name: type
          required: false
          description: Optional transaction type filter.
          schema:
            type: string
            enum: [buy, sell, swap, deposit, withdrawal]
          example: buy
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
          example: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
          example: 20
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
          example: desc
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/TransactionListResponse"
  /wallet/transactions/{transactionId}:
    get:
      tags: [Wallet]
      summary: Get transaction details
      parameters:
        - in: path
          name: transactionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/TransactionResponse"
  /wallet/deposit/simulate:
    post:
      tags: [Wallet]
      summary: Simulate a sandbox USD deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  example: 500
      responses:
        "201":
          description: Deposit transaction
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/DepositResponse"
  /wallet/withdrawals:
    post:
      tags: [Wallet]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetSymbol, amount, address, network]
              properties:
                assetSymbol:
                  type: string
                  example: ETH
                amount:
                  type: number
                  example: 0.05
                address:
                  type: string
                network:
                  type: string
      responses:
        "201":
          description: Withdrawal request
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/WithdrawalResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Symbol:
      in: path
      name: symbol
      required: true
      schema:
        type: string
      example: BTC
  examples:
...
    DepositAddressResponse:
      summary: Deposit address
      value:
        data:
          assetSymbol: USDC
          network: Base Sepolia
          address: "0x3333333333333333333333333333333333333333"
          qrPayload: "ethereum:0x3333333333333333333333333333333333333333@84532"
    TransactionResponse:
      summary: Transaction detail
      value:
        data:
          id: txn_swap_001
          userId: usr_student
          type: swap
          status: completed
          fromAsset: ETH
          toAsset: USDC
          fromAmount: 0.1
          toAmount: 306.12
          feeAmount: 2.48
          rate: 3100.25
          reference: CRT-SWAP-1780000000000
          note: Sandbox trade execution
          createdAt: "2026-05-03T14:12:00.000Z"
          completedAt: "2026-05-03T14:12:00.000Z"
    TransactionListResponse:
      summary: Transaction list
      value:
        data:
          - id: txn_swap_001
            userId: usr_student
            type: swap
            status: completed
            fromAsset: ETH
            toAsset: USDC
            fromAmount: 0.1
            toAmount: 306.12
            feeAmount: 2.48
            rate: 3100.25
            reference: CRT-SWAP-1780000000000
            note: Sandbox trade execution
            createdAt: "2026-05-03T14:12:00.000Z"
            completedAt: "2026-05-03T14:12:00.000Z"
        meta:
          count: 1
    DepositResponse:
      summary: Sandbox deposit
      value:
        data:
          transaction:
            id: txn_deposit_001
            userId: usr_student
            type: deposit
            status: completed
            fromAsset: USD
            toAsset: USD
            fromAmount: 500
            toAmount: 500
            feeAmount: 0
            rate: 1
            reference: CRT-DEP-1780000000000
            note: Sandbox deposit
            createdAt: "2026-05-03T14:10:00.000Z"
            completedAt: "2026-05-03T14:10:00.000Z"
          wallet:
            id: wallet_student
            userId: usr_student
            fiatCurrency: USD
            depositAddresses: []
            balances:
              - assetSymbol: USD
                available: 1500
                locked: 0
    WithdrawalResponse:
      summary: Withdrawal request
      value:
        data:
          id: wd_001
          userId: usr_student
          assetSymbol: ETH
          amount: 0.05
          feeAssetAmount: 0.00064511
          address: "0xabc123abc123abc123abc123abc123abc123abcd"
          network: Ethereum Sepolia
          status: pending
          createdAt: "2026-05-03T14:20:00.000Z"
          reviewedAt: null
          reviewerNote: null
...